removed termination table, started making token eater

+++ dbl-0.2/src/random/random_float.c
@@ -4,5 +4,5 @@
 
 float random_float(float min, float max)
 {
-    return min + (float) rand() / RAND_MAX * (max - min);
+    return min + (float) rand() / (float) RAND_MAX * (max - min);
 }
